cleanup event definitions
