+++ src/sentry/analytics/events/ai_autofix_pr_events.py
@@ -1,0 +1,30 @@
+from sentry import analytics
+
+
+class AiAutofixPrEvent(analytics.Event):
+    type = "ai.autofix.pr.closed"
+
+    attributes = (
+        analytics.Attribute("organization_id"),
+        analytics.Attribute("project_id"),
+        analytics.Attribute("group_id"),
+        analytics.Attribute("run_id"),
+        analytics.Attribute("integration"),
+    )
+
+
+class AiAutofixPrClosedEvent(AiAutofixPrEvent):
+    type = "ai.autofix.pr.closed"
+
+
+class AiAutofixPrMergedEvent(analytics.Event):
+    type = "ai.autofix.pr.merged"
+
+
+class AiAutofixPrOpenedEvent(AiAutofixPrEvent):
+    type = "ai.autofix.pr.opened"
+
+
+analytics.register(AiAutofixPrClosedEvent)
+analytics.register(AiAutofixPrMergedEvent)
+analytics.register(AiAutofixPrOpenedEvent)
