--- src/sentry/analytics/events/ai_autofix_pr_opened.py
@@ -1,16 +1,0 @@
-from sentry import analytics
-
-
-class AiAutofixPrOpenedEvent(analytics.Event):
-    type = "ai.autofix.pr.opened"
-
-    attributes = (
-        analytics.Attribute("organization_id"),
-        analytics.Attribute("project_id"),
-        analytics.Attribute("group_id"),
-        analytics.Attribute("run_id"),
-        analytics.Attribute("integration"),
-    )
-
-
-analytics.register(AiAutofixPrOpenedEvent)
